+++ test/initial_sync_repro_test.dart
@@ -31,6 +31,9 @@ class FakeGanttSyncClient implements GanttSyncClient {
 
   @override
   Future<void> syncWithMerkle({required String remoteRoot, required int depth}) async {}
+
+  @override
+  String get actorId => 'test-actor';
 }
 
 void main() {