+++ test/src/legacy_gantt_behavior_test.dart
@@ -468,4 +468,7 @@ class MockGanttSyncClient extends GanttSyncClient {
 
   @override
   Future<void> syncWithMerkle({required String remoteRoot, required int depth}) async {}
+
+  @override
+  String get actorId => 'mock-actor';
 }