+++ test/src/legacy_gantt_view_model_cursor_test.dart
@@ -48,6 +48,9 @@ class MockGanttSyncClient extends GanttSyncClient {
 
   @override
   Future<void> syncWithMerkle({required String remoteRoot, required int depth}) async {}
+
+  @override
+  String get actorId => 'mock-actor';
 }
 
 void main() {