+++ test/src/legacy_gantt_view_model_test.dart
@@ -900,4 +900,7 @@ class MockGanttSyncClient extends GanttSyncClient {
 
   @override
   Future<void> syncWithMerkle({required String remoteRoot, required int depth}) async {}
+
+  @override
+  String get actorId => 'mock-actor';
 }